+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2021-09-07
+
 ### Added
 
 - Fuzzing case for parsing of a `Task`
@@ -48,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic parsing capabilities
 
-[Unreleased]: https://github.com/Shemnei/tdtxt/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/Shemnei/tdtxt/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/Shemnei/tdtxt/releases/tag/v0.3.0
 [0.2.0]: https://github.com/Shemnei/tdtxt/releases/tag/v0.2.0
 [0.1.0]: https://github.com/Shemnei/tdtxt/releases/tag/v0.1.0